Update of the backend and fully finnished

+++ individual-assignmet-db/src/main/java/com/example/accessingdatamysql/Controllers/CategoryController.java
@@ -3,7 +3,10 @@ package com.example.accessingdatamysql.Controllers;
 import com.example.accessingdatamysql.Repository.CategoryRepository;
 import com.example.accessingdatamysql.ResourceNotFoundException;
 import com.example.accessingdatamysql.modelsTemp.Category;
+import com.example.accessingdatamysql.modelsTemp.Post;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,18 +16,11 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 @RequestMapping(path = "/category")
-@CrossOrigin(origins = {"http://localhost:3001", "http://localhost:3000"})
+@CrossOrigin(origins = "http://localhost:3000")
 
 public class CategoryController {
     @Autowired
     private CategoryRepository categoryRepository;
-
-    @PostMapping(path = "/add")
-    public @ResponseBody
-    Category addNewCategory(@RequestBody Category category) {
-        categoryRepository.save(category);
-        return category;
-    }
 
     @GetMapping(path = "/all")
     public @ResponseBody
@@ -32,16 +28,33 @@ public class CategoryController {
         return categoryRepository.findAll();
     }
 
+    @PostMapping("/add")
+    public ResponseEntity<Category> addNewCategory(@RequestBody Category category) {
+        categoryRepository.save(category);
+        return new ResponseEntity<Category>(HttpStatus.NOT_FOUND);
+    }
+
     @PutMapping("/{id}/update")
-    public Category updateCategory(@PathVariable long categoryId, @RequestBody Category categoryRequest) {
-        return categoryRepository.findById(categoryId).map(c -> {
-            c.setName(categoryRequest.getName());
-            return categoryRepository.save(c);
-        }).orElseThrow(() -> new ResourceNotFoundException("CategoryId " + categoryId + " not found"));
+    public ResponseEntity<Category> updateCategory(@PathVariable long categoryId, @RequestBody Category updatedCategory) {
+        if (categoryRepository.existsById(categoryId)){
+            categoryRepository.findById(categoryId).map(c -> {
+                c.setName(c.getName());
+                categoryRepository.save(c);
+                return updatedCategory;
+            });
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        else
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("/{id}/delete")
-    public void deleteCategory(@PathVariable("id") long id) throws Exception {
-        categoryRepository.deleteById(id);
+    public ResponseEntity<Category> deleteCategory(@PathVariable long id){
+        if (categoryRepository.existsById(id)) {
+            categoryRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
+
 }
