+++ individual-assignmet-db/src/main/java/com/example/accessingdatamysql/Controllers/CommentController.java
@@ -4,15 +4,17 @@ import com.example.accessingdatamysql.Repository.PostRepository;
 import com.example.accessingdatamysql.ResourceNotFoundException;
 import com.example.accessingdatamysql.modelsTemp.Comment;
 import com.example.accessingdatamysql.modelsTemp.Product;
+import com.example.accessingdatamysql.modelsTemp.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 
 @RestController
 @RequestMapping(path = "/comment")
-@CrossOrigin(origins = {"http://localhost:3001", "http://localhost:3000"})
+@CrossOrigin(origins = "http://localhost:3000")
 
 public class CommentController {
 
@@ -35,28 +37,33 @@ public class CommentController {
             return commentRepository.save(comment);
         }).orElseThrow(() -> new ResourceNotFoundException("PostId " + postId + " not found"));
     }
-    @PostMapping(path = "/add")
-    public @ResponseBody
-    Comment addNewComment(@RequestBody Comment comment) {
-        commentRepository.save(comment);
-        return comment;
-    }
-    @PutMapping("/posts/{postId}/comments/{commentId}")
-    public Comment updateComment(@PathVariable (value = "postId") Long postId,
-                                 @PathVariable (value = "commentId") Long commentId,
-                                 @RequestBody Comment commentRequest) {
-        if(!postRepository.existsById(postId)) {
-            throw new ResourceNotFoundException("PostId " + postId + " not found");
-        }
 
-        return commentRepository.findById(commentId).map(comment -> {
-            comment.setText(commentRequest.getText());
-            return commentRepository.save(comment);
-        }).orElseThrow(() -> new ResourceNotFoundException("CommentId " + commentId + "not found"));
+    @PostMapping("/add")
+    public ResponseEntity<Comment> addNewComment(@RequestBody Comment comment) {
+        commentRepository.save(comment);
+        return new ResponseEntity<Comment>(HttpStatus.NOT_FOUND);
+    }
+
+    @PutMapping("/posts/{postId}/comments/{commentId}/update")
+    public ResponseEntity<Comment> updateComment(@PathVariable long commentid,@PathVariable long postId, @RequestBody Comment updatedComment) {
+        if (postRepository.existsById(postId)){
+            commentRepository.findById(commentid).map(c -> {
+                c.setText(c.getText());
+                commentRepository.save(c);
+                return updatedComment;
+            });
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        else
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("/{id}/delete")
-    public void deleteComment(@PathVariable("id") long id) throws Exception {
-        commentRepository.deleteById(id);
+    public ResponseEntity<Comment> deleteComment(@PathVariable long id){
+        if (commentRepository.existsById(id)) {
+            commentRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
