+++ individual-assignmet-db/src/main/java/com/example/accessingdatamysql/Controllers/ProductController.java
@@ -5,6 +5,8 @@ import com.example.accessingdatamysql.ResourceNotFoundException;
 import com.example.accessingdatamysql.modelsTemp.Post;
 import com.example.accessingdatamysql.modelsTemp.Product;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,35 +16,46 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 @RequestMapping(path = "/product")
-@CrossOrigin(origins = {"http://localhost:3001", "http://localhost:3000"})
+@CrossOrigin(origins = "http://localhost:3000")
 
 public class ProductController {
     @Autowired
     private ProductRepository productRepository;
-
-    @PostMapping(path = "/add")
-    public @ResponseBody Product addNewProduct(@RequestBody Product product) {
-        productRepository.save(product);
-        return product;
-    }
 
     @GetMapping(path = "/all")
     public @ResponseBody
     Iterable<Product> getAllProducts() {
         return productRepository.findAll();
     }
-    
-    @PutMapping("/{id}/update")
-    public Product updateProduct(@PathVariable long productId, @RequestBody Product productRequest) {
-        return productRepository.findById(productId).map(product -> {
-            product.setProductName(productRequest.getProductName());
-            product.setPrice(productRequest.getPrice());
-            product.setCategory(productRequest.getCategory());
-            return productRepository.save(product);
-        }).orElseThrow(() -> new ResourceNotFoundException("ProductId " + productId + " not found"));
+
+    @PostMapping("/add")
+    public ResponseEntity<Product> addNewProduct(@RequestBody Product product) {
+        productRepository.save(product);
+        return new ResponseEntity<Product>(HttpStatus.NOT_FOUND);
     }
+
+
+    @PutMapping("/{id}/update")
+    public ResponseEntity<Product> updateProduct(@PathVariable long productId, @RequestBody Product updatedProduct) {
+        if (productRepository.existsById(productId)){
+            productRepository.findById(productId).map(p -> {
+                p.setProductName(p.getProductName());
+                p.setPrice(p.getPrice());
+                productRepository.save(p);
+                return updatedProduct;
+            });
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        else
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
     @DeleteMapping("/{id}/delete")
-    public void deleteUser(@PathVariable("id") long id) throws Exception {
-        productRepository.deleteById(id);
+    public ResponseEntity<Product> deleteProduct(@PathVariable long id){
+        if (productRepository.existsById(id)) {
+            productRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
